Fix CI: Consolidate environment config into Supabase.swift
Move SupabaseEnvironment and SupabaseConfiguration into Supabase.swift
to avoid Xcode project file manipulation. The separate file was causing
CI build failures because it wasn't properly added to Xcode project targets.

Changes:
- Consolidate all environment logic into Spot/Supabase/Supabase.swift
- Delete Spot/Utils/SupabaseEnvironment.swift
- Functionality identical (still uses #if DEBUG for env selection)
- Fixes exit code 65 build failure in PR validation
